move types to new file
